# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation - Basic CRUD Laravel",
        "contact": {
            "email": "listcreative15@gmail.com"
        },
        "license": {
            "name": "GPL2",
            "url": "https://devsenv.com"
        },
        "version": "1.0.1"
    },
  "servers": [
    {
      "url": "/"
    }
  ],
    "paths": {
...
      "/api/products/{id}": {
        "get": {
          "tags": [
            "Products"
          ],
          "summary": "Show Product Details",
          "description": "Show Product Details",
          "operationId": "show",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id, eg; 1",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Show Product Details"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Resource Not Found"
            }
          },
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "put": {
          "tags": [
            "Products"
          ],
          "summary": "Update Product",
          "description": "Update Product",
          "operationId": "update",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id, eg; 1",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/products_id_body"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Update Product"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Resource Not Found"
            }
          },
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "delete": {
          "tags": [
            "Products"
          ],
          "summary": "Delete Product",
          "description": "Delete Product",
          "operationId": "destroy",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id, eg; 1",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Delete Product"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Resource Not Found"
            }
          },
          "security": [
            {
              "bearer": []
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
...
        "products_id_body": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "example": "Product 1"
            },
            "description": {
              "type": "string",
              "example": "Description"
            },
            "price": {
              "type": "integer",
              "example": 10120
            },
            "image": {
              "type": "string",
              "example": ""
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearer": {
          "type": "apiKey",
          "description": "Bearer token, Ex: Bearer <token>",
          "name": "Authorization",
          "in": "header",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}